test(cli): split config option tests
